Move and register interpreter path command
